%% File Name: STARTS-package.Rd
%% File Version: 0.12

\name{STARTS-package}
\alias{STARTS-package}
\alias{STARTS}
\docType{package}

\title{
\packageTitle{STARTS} 
}

\description{
\packageDescription{STARTS}
}

\author{
\packageAuthor{STARTS}

Maintainer: \packageMaintainer{STARTS}
}

%% include references
\references{
Kenny, D. A., & Zautra, A. (1995). The trait-state-error 
model for multiwave data. 
\emph{Journal of Consulting and Clinical Psychology, 63}, 52-59.

Kenny, D. A., & Zautra, A. (2001). Trait-state models for 
longitudinal data. In L. M. Collins & A. G. Sayer (Eds.), 
\emph{New methods for the analysis of change} (pp. 243-263).
Washington, DC, US: American Psychological Association.

Luedtke, O., Robitzsch, A., & Wagner, J. (2018). More stable estimation of the STARTS model:  
A Bayesian approach using Markov Chain Monte Carlo techniques. 
\emph{Psychological Methods, 23}(3), 570-593.
}

%\seealso{
%XXXXXXX
%}

\keyword{package}
