%% File Name: STARTS-utilities.Rd
%% File Version: 0.03

\name{STARTS-utilities}
\alias{digamma2}
%- Also NEED an '\alias' for EACH other topic documented here.



\title{
Utility Functions in the \pkg{STARTS} Package
}


\description{
Utlity functions in the \pkg{STARTS} package
}

\usage{
## density inverse gamma distribution
digamma2(x, n0, var0)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Numeric Vector
}
  \item{n0}{
Prior sample size
}
  \item{var0}{
Prior variance
}
}

%%   'object' 'digits' 'file' '...'

%%\details{
%% xxx
%% }




%\examples{
%library(sirt)
%}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{vvvvv}
%\keyword{vvvvv}% __ONLY ONE__ keyword per line
