\name{STARTS-package}
\alias{STARTS-package}
\alias{STARTS}
\docType{package}

\title{
Functions for the STARTS model
}

\description{
    Contains functions for estimating the STARTS model of
    Kenny and Zautra (1995, 2001) <DOI:10.1037/0022-006X.63.1.52>,
    <DOI:10.1037/10409-008>.
}

%% \details{

%% The \pkg{LAM} package contains the following main functions: 

%% \itemize{   
%%
%% mlnormal
%% \item A general fitt
%%
%% amh, pmle
%%\item A general (but
%%
%% }

% }


\author{
Alexander Robitzsch, Oliver Luedtke
IPN - Leibniz Institute for Science and Mathematics Education
at Kiel University \cr \cr
Maintainer: Alexander Robitzsch <robitzsch@ipn.uni-kiel.de> \cr \cr
}


%% include references
\references{
Kenny, D. A., & Zautra, A. (1995). The trait-state-error 
model for multiwave data. 
\emph{Journal of Consulting and Clinical Psychology}, 
\bold{63}, 52-59.

Kenny, D. A., & Zautra, A. (2001). Trait-state models for 
longitudinal data. In L. M. Collins & A. G. Sayer (Eds.), 
\emph{New methods for the analysis of change} (pp. 243-263).
Washington, DC, US: American Psychological Association.

}

%\seealso{
%XXXXXXX
%}

\keyword{package}
